wowowow trop bien

+++ src/syntax/token/op/And.java
@@ -1,10 +1,10 @@
 package syntax.token.op;
 
 import syntax.token.Token;
-import syntax.token.Value;
+import syntax.token.val.Value;
 
 public class And extends BinaryOperation<Object, Object, Double> {
     public And(Token[] tokens) {
-        super((a, b) -> (double) (Value.isTruthy(a) && Value.isTruthy(b) ? 1 : 0), tokens);
+        super((a, b) -> (Value.isTruthy(a) && Value.isTruthy(b) ? 1d : 0d), tokens);
     }
 }
