+++ src/syntax/token/op/Not.java
@@ -1,10 +1,10 @@
 package syntax.token.op;
 
 import syntax.token.Token;
-import syntax.token.Value;
+import syntax.token.val.Value;
 
 public class Not extends UnaryOperation<Object, Double> {
     public Not(Token[] tokens) {
-        super((a) -> (double) (Value.isTruthy(a) ? 0 : 1), tokens);
+        super((a) -> (Value.isTruthy(a) ? 0d : 1d), tokens);
     }
 }
