+++ src/syntax/token/op/Nullity.java
@@ -4,6 +4,6 @@ import syntax.token.Token;
 
 public class Nullity extends UnaryOperation<Object, Double> {
     public Nullity(Token[] tokens) {
-        super(a -> (double) (a != null ? 1 : 0), tokens);
+        super(a -> (a != null ? 1d : 0d), tokens);
     }
 }
